API polishing up.

+++ src/Controller/TaskApiController.php
@@ -40,12 +40,20 @@ class TaskApiController extends AbstractController
 
         $task->setTitle($request->get('title'));
         $task->setComment($request->get('comment'));
+        $task->setTimeSpent(
+            \DateInterval::createFromDateString($request->get('time-spend'))
+        );
+        $task->setDateStarted(
+            new \DateTime($request->get('date-started'))
+        );
+
         $em->persist($task);
         $em->flush();
 
         return $this->json($task);
     }
 
+    #[Route('/api/v1/task/{id}', methods: ['DELETE'])]
     /**
      * Deletion
      *
@@ -53,16 +61,47 @@ class TaskApiController extends AbstractController
      * 
      * @return Response
      */
-    #[Route('/api/v1/task/{id}', methods: ['DELETE'])]
     public function delete(int $id): Response
     {
         $em = $this->getDoctrine()->getManager();
-        
+
         $taskRepository = $em->getRepository(Task::class);
         $task = $taskRepository->find($id);
+
+        if (!$task) {
+            throw new NotFoundHttpException();
+        }
+        
         $em->remove($task);
         $em->flush();
 
         return $this->json('');
+    }
+
+    #[Route('/api/v1/task', methods: ['POST'])]
+    /**
+     * Task creation
+     *
+     * @param Request $request Request
+     * 
+     * @return JsonResponse
+     */
+    public function create(Request $request): JsonResponse
+    {
+        $task = new Task();
+        $task->setTitle($request->request->get('title'));
+        $task->setComment($request->request->get('comment'));
+        $task->setTimeSpent(
+            \DateInterval::createFromDateString(
+                $request->request->get('time-spent')
+            )
+        );
+        $task->setDateStarted(new \DateTime($request->request->get('date-started')));
+
+        $em = $this->getDoctrine()->getManager();
+        $em->persist($task);
+        $em->flush();
+
+        return $this->json($task);
     }
 }
